third_party/WebKit/LayoutTests/fast/css/invalidation/shadow-add-sheet-host-expected.txt

CONSOLE WARNING: line 14: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Insert a style element into a shadow tree affecting the host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(host2).color is "rgb(255, 0, 0)"
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(host2).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
